reorganising sample logging into new script
