+++ website/__init__.py
@@ -46,8 +46,10 @@ def create_app():
 
     from .views import views
     from .registrations import registrations
+    from .logsamples import logsamples
 
     app.register_blueprint(views, url_prefix='/')
     app.register_blueprint(registrations, url_prefix='/')
+    app.register_blueprint(logsamples, url_prefix='/')
 
     return app
